Add files via upload
This time, I had the opportunity to dive deep into the Titanic dataset, where I was assigned to perform Exploratory Data Analysis (EDA) and elaborate on the scenario drawn from the data.

🔍 I started by cleaning and transforming the dataset to ensure it was ready for analysis. Following that, I conducted EDA to uncover key insights and patterns within the data. To visualize these findings, I created various plots, including a heatmap that effectively showcased the relationships between different variables and gained insights .

This task was a great learning experience, reinforcing my skills in data cleaning, transformation, and visualization. I'm excited to continue growing and tackling more challenges in my journey as a data scientist!